+++ ui-weather-view/src/main/kotlin/org/breezyweather/ui/theme/weatherView/materialWeatherView/WeatherImplementorFactory.kt
@@ -47,7 +47,8 @@ object WeatherImplementorFactory {
             WeatherView.WEATHER_KIND_CLEAR -> if (daytime) {
                 SunImplementor(
                     sizes,
-                    animate
+                    animate,
+                    darkMode
                 )
             } else {
                 MeteorShowerImplementor(