+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 **Sources**
 - Met Office (UK) - Fix shifted day sequence in locations with a different timezone than UTC
+- NCEI - Fix parsing error
 
 
 # Version 6.2.1 (2026-06-07)